refactor: moved requests and responses to separate folders
