+++ driver/src/main/java/com/modsen/driver/dto/request/DriverCreate.java
@@ -1,7 +1,8 @@
-package com.modsen.driver.dto;
+package com.modsen.driver.dto.request;
 
-import com.modsen.driver.model.Status;
 import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Pattern;
 import jakarta.validation.constraints.Size;
 import lombok.AllArgsConstructor;
@@ -11,19 +12,26 @@ import lombok.NoArgsConstructor;
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class DriverUpdate {
+public class DriverCreate {
 
+    @NotNull(message = "Name must not be null")
+    @NotBlank(message = "Name must not be blank")
     @Size(min = 2, max = 255, message = "Name size must be between 2 and 255")
     private String name;
 
+    @NotNull(message = "Surname must not be null")
+    @NotBlank(message = "Surname must not be blank")
     @Size(min = 2, max = 255, message = "Surname size must be between 2 and 255")
     private String surname;
 
+    @NotNull(message = "Email must not be null")
+    @NotBlank(message = "Email must not be blank")
     @Email(regexp = "^[A-Za-z0-9._%+-]+@[A-Za-z0-9-]+\\.[A-Za-z]{2,}$", message = "Invalid email format")
     private String email;
 
+    @NotNull(message = "Phone number must not be null")
+    @NotBlank(message = "Phone number must not be blank")
     @Size(min = 10, max = 15, message = "Phone number must be between 10 and 15 characters")
     @Pattern(regexp = "\\d+", message = "Phone number must contain only digits")
     private String phoneNumber;
-    private Status status;
 }
