+++ driver/src/main/java/com/modsen/driver/dto/response/ErrorResponse.java
@@ -1,16 +1,14 @@
-package com.modsen.driver.dto;
+package com.modsen.driver.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.util.List;
-
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class ValidationErrorResponse {
+public class ErrorResponse {
 
     private Integer statusCode;
-    private List<String> errors;
+    private String errorMessage;
 }
